+++ cpu/ctensorapplyfunc.h
@@ -268,11 +268,11 @@ void CTensorApplyFunc<type_t>::ApplySigmoid(CTensor<type_t> &cTensor_Output,cons
  const type_t *input_ptr=&cTensor_Input.Item[0];
  type_t *o_ptr=&cTensor_Output.Item[0];
 
- for(size_t z=0;z<cTensor_Input.Size_Z;z++)
+ for(uint32_t z=0;z<cTensor_Input.Size_Z;z++)
  {
-  for(size_t y=0;y<cTensor_Input.Size_Y;y++)
+  for(uint32_t y=0;y<cTensor_Input.Size_Y;y++)
   {
-   for(size_t x=0;x<cTensor_Input.Size_X;x++,o_ptr++,input_ptr++)
+   for(uint32_t x=0;x<cTensor_Input.Size_X;x++,o_ptr++,input_ptr++)
    {
     *o_ptr=Sigmoid(*input_ptr);
    }
@@ -294,11 +294,11 @@ void CTensorApplyFunc<type_t>::ApplyReLU(CTensor<type_t> &cTensor_Output,const C
  const type_t *input_ptr=&cTensor_Input.Item[0];
  type_t *o_ptr=&cTensor_Output.Item[0];
 
- for(size_t z=0;z<cTensor_Input.Size_Z;z++)
+ for(uint32_t z=0;z<cTensor_Input.Size_Z;z++)
  {
-  for(size_t y=0;y<cTensor_Input.Size_Y;y++)
+  for(uint32_t y=0;y<cTensor_Input.Size_Y;y++)
   {
-   for(size_t x=0;x<cTensor_Input.Size_X;x++,o_ptr++,input_ptr++)
+   for(uint32_t x=0;x<cTensor_Input.Size_X;x++,o_ptr++,input_ptr++)
    {
     *o_ptr=ReLU(*input_ptr);
    }
@@ -320,11 +320,11 @@ void CTensorApplyFunc<type_t>::ApplyGeLU(CTensor<type_t> &cTensor_Output,const C
  const type_t *input_ptr=&cTensor_Input.Item[0];
  type_t *o_ptr=&cTensor_Output.Item[0];
 
- for(size_t z=0;z<cTensor_Input.Size_Z;z++)
+ for(uint32_t z=0;z<cTensor_Input.Size_Z;z++)
  {
-  for(size_t y=0;y<cTensor_Input.Size_Y;y++)
+  for(uint32_t y=0;y<cTensor_Input.Size_Y;y++)
   {
-   for(size_t x=0;x<cTensor_Input.Size_X;x++,o_ptr++,input_ptr++)
+   for(uint32_t x=0;x<cTensor_Input.Size_X;x++,o_ptr++,input_ptr++)
    {
     *o_ptr=GeLU(*input_ptr);
    }
@@ -346,11 +346,11 @@ void CTensorApplyFunc<type_t>::ApplyLeakyReLU(CTensor<type_t> &cTensor_Output,co
  const type_t *input_ptr=&cTensor_Input.Item[0];
  type_t *o_ptr=&cTensor_Output.Item[0];
 
- for(size_t z=0;z<cTensor_Input.Size_Z;z++)
+ for(uint32_t z=0;z<cTensor_Input.Size_Z;z++)
  {
-  for(size_t y=0;y<cTensor_Input.Size_Y;y++)
+  for(uint32_t y=0;y<cTensor_Input.Size_Y;y++)
   {
-   for(size_t x=0;x<cTensor_Input.Size_X;x++,o_ptr++,input_ptr++)
+   for(uint32_t x=0;x<cTensor_Input.Size_X;x++,o_ptr++,input_ptr++)
    {
     *o_ptr=LeakyReLU(*input_ptr);
    }
@@ -372,11 +372,11 @@ void CTensorApplyFunc<type_t>::ApplyLinear(CTensor<type_t> &cTensor_Output,const
  const type_t *input_ptr=&cTensor_Input.Item[0];
  type_t *o_ptr=&cTensor_Output.Item[0];
 
- for(size_t z=0;z<cTensor_Input.Size_Z;z++)
+ for(uint32_t z=0;z<cTensor_Input.Size_Z;z++)
  {
-  for(size_t y=0;y<cTensor_Input.Size_Y;y++)
+  for(uint32_t y=0;y<cTensor_Input.Size_Y;y++)
   {
-   for(size_t x=0;x<cTensor_Input.Size_X;x++,o_ptr++,input_ptr++)
+   for(uint32_t x=0;x<cTensor_Input.Size_X;x++,o_ptr++,input_ptr++)
    {
     *o_ptr=Linear(*input_ptr);
    }
@@ -398,11 +398,11 @@ void CTensorApplyFunc<type_t>::ApplyTangence(CTensor<type_t> &cTensor_Output,con
  const type_t *input_ptr=&cTensor_Input.Item[0];
  type_t *o_ptr=&cTensor_Output.Item[0];
 
- for(size_t z=0;z<cTensor_Input.Size_Z;z++)
+ for(uint32_t z=0;z<cTensor_Input.Size_Z;z++)
  {
-  for(size_t y=0;y<cTensor_Input.Size_Y;y++)
+  for(uint32_t y=0;y<cTensor_Input.Size_Y;y++)
   {
-   for(size_t x=0;x<cTensor_Input.Size_X;x++,o_ptr++,input_ptr++)
+   for(uint32_t x=0;x<cTensor_Input.Size_X;x++,o_ptr++,input_ptr++)
    {
     *o_ptr=Tangence(*input_ptr);
    }
@@ -424,11 +424,11 @@ void CTensorApplyFunc<type_t>::ApplySoftMax(CTensor<type_t> &cTensor_Output,cons
  const type_t *input_ptr=&cTensor_Input.Item[0];
  type_t *o_ptr=&cTensor_Output.Item[0];
 
- for(size_t z=0;z<cTensor_Input.Size_Z;z++)
+ for(uint32_t z=0;z<cTensor_Input.Size_Z;z++)
  {
-  for(size_t y=0;y<cTensor_Input.Size_Y;y++)
+  for(uint32_t y=0;y<cTensor_Input.Size_Y;y++)
   {
-   for(size_t x=0;x<cTensor_Input.Size_X;x++,o_ptr++,input_ptr++)
+   for(uint32_t x=0;x<cTensor_Input.Size_X;x++,o_ptr++,input_ptr++)
    {
     *o_ptr=SoftMax(*input_ptr);
    }
@@ -451,11 +451,11 @@ void CTensorApplyFunc<type_t>::ApplyDifferentialSigmoid(CTensor<type_t> &cTensor
  const type_t *input_ptr=&cTensor_Input.Item[0];
  type_t *o_ptr=&cTensor_Output.Item[0];
 
- for(size_t z=0;z<cTensor_Input.Size_Z;z++)
+ for(uint32_t z=0;z<cTensor_Input.Size_Z;z++)
  {
-  for(size_t y=0;y<cTensor_Input.Size_Y;y++)
+  for(uint32_t y=0;y<cTensor_Input.Size_Y;y++)
   {
-   for(size_t x=0;x<cTensor_Input.Size_X;x++,o_ptr++,input_ptr++)
+   for(uint32_t x=0;x<cTensor_Input.Size_X;x++,o_ptr++,input_ptr++)
    {
     *o_ptr=dSigmoid(*input_ptr);
    }
@@ -476,11 +476,11 @@ void CTensorApplyFunc<type_t>::ApplyDifferentialReLU(CTensor<type_t> &cTensor_Ou
  const type_t *input_ptr=&cTensor_Input.Item[0];
  type_t *o_ptr=&cTensor_Output.Item[0];
 
- for(size_t z=0;z<cTensor_Input.Size_Z;z++)
+ for(uint32_t z=0;z<cTensor_Input.Size_Z;z++)
  {
-  for(size_t y=0;y<cTensor_Input.Size_Y;y++)
+  for(uint32_t y=0;y<cTensor_Input.Size_Y;y++)
   {
-   for(size_t x=0;x<cTensor_Input.Size_X;x++,o_ptr++,input_ptr++)
+   for(uint32_t x=0;x<cTensor_Input.Size_X;x++,o_ptr++,input_ptr++)
    {
     *o_ptr=dReLU(*input_ptr);
    }
@@ -502,11 +502,11 @@ void CTensorApplyFunc<type_t>::ApplyDifferentialGeLU(CTensor<type_t> &cTensor_Ou
  const type_t *input_ptr=&cTensor_Input.Item[0];
  type_t *o_ptr=&cTensor_Output.Item[0];
 
- for(size_t z=0;z<cTensor_Input.Size_Z;z++)
+ for(uint32_t z=0;z<cTensor_Input.Size_Z;z++)
  {
-  for(size_t y=0;y<cTensor_Input.Size_Y;y++)
+  for(uint32_t y=0;y<cTensor_Input.Size_Y;y++)
   {
-   for(size_t x=0;x<cTensor_Input.Size_X;x++,o_ptr++,input_ptr++)
+   for(uint32_t x=0;x<cTensor_Input.Size_X;x++,o_ptr++,input_ptr++)
    {
     *o_ptr=dGeLU(*input_ptr);
    }
@@ -528,11 +528,11 @@ void CTensorApplyFunc<type_t>::ApplyDifferentialLeakyReLU(CTensor<type_t> &cTens
  const type_t *input_ptr=&cTensor_Input.Item[0];
  type_t *o_ptr=&cTensor_Output.Item[0];
 
- for(size_t z=0;z<cTensor_Input.Size_Z;z++)
+ for(uint32_t z=0;z<cTensor_Input.Size_Z;z++)
  {
-  for(size_t y=0;y<cTensor_Input.Size_Y;y++)
+  for(uint32_t y=0;y<cTensor_Input.Size_Y;y++)
   {
-   for(size_t x=0;x<cTensor_Input.Size_X;x++,o_ptr++,input_ptr++)
+   for(uint32_t x=0;x<cTensor_Input.Size_X;x++,o_ptr++,input_ptr++)
    {
     *o_ptr=dLeakyReLU(*input_ptr);
    }
@@ -553,11 +553,11 @@ void CTensorApplyFunc<type_t>::ApplyDifferentialLinear(CTensor<type_t> &cTensor_
  const type_t *input_ptr=&cTensor_Input.Item[0];
  type_t *o_ptr=&cTensor_Output.Item[0];
 
- for(size_t z=0;z<cTensor_Input.Size_Z;z++)
+ for(uint32_t z=0;z<cTensor_Input.Size_Z;z++)
  {
-  for(size_t y=0;y<cTensor_Input.Size_Y;y++)
+  for(uint32_t y=0;y<cTensor_Input.Size_Y;y++)
   {
-   for(size_t x=0;x<cTensor_Input.Size_X;x++,o_ptr++,input_ptr++)
+   for(uint32_t x=0;x<cTensor_Input.Size_X;x++,o_ptr++,input_ptr++)
    {
     *o_ptr=dLinear(*input_ptr);
    }
@@ -578,11 +578,11 @@ void CTensorApplyFunc<type_t>::ApplyDifferentialTangence(CTensor<type_t> &cTenso
  const type_t *input_ptr=&cTensor_Input.Item[0];
  type_t *o_ptr=&cTensor_Output.Item[0];
 
- for(size_t z=0;z<cTensor_Input.Size_Z;z++)
+ for(uint32_t z=0;z<cTensor_Input.Size_Z;z++)
  {
-  for(size_t y=0;y<cTensor_Input.Size_Y;y++)
+  for(uint32_t y=0;y<cTensor_Input.Size_Y;y++)
   {
-   for(size_t x=0;x<cTensor_Input.Size_X;x++,o_ptr++,input_ptr++)
+   for(uint32_t x=0;x<cTensor_Input.Size_X;x++,o_ptr++,input_ptr++)
    {
     *o_ptr=dTangence(*input_ptr);
    }
@@ -604,11 +604,11 @@ void CTensorApplyFunc<type_t>::ApplyDifferentialSoftMax(CTensor<type_t> &cTensor
  const type_t *input_ptr=&cTensor_Input.Item[0];
  type_t *o_ptr=&cTensor_Output.Item[0];
 
- for(size_t z=0;z<cTensor_Input.Size_Z;z++)
+ for(uint32_t z=0;z<cTensor_Input.Size_Z;z++)
  {
-  for(size_t y=0;y<cTensor_Input.Size_Y;y++)
+  for(uint32_t y=0;y<cTensor_Input.Size_Y;y++)
   {
-   for(size_t x=0;x<cTensor_Input.Size_X;x++,o_ptr++,input_ptr++)
+   for(uint32_t x=0;x<cTensor_Input.Size_X;x++,o_ptr++,input_ptr++)
    {
     *o_ptr=dSoftMax(*input_ptr);
    }
